requirements.txt
numpy==2.0.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
openai==1.55.3
scikit-learn==1.5.2
pydantic==2.10.6
tabulate==0.9.0
pytest==8.3.5